+++ suppaftp/src/async_ftp/tokio_ftp/tls.rs
@@ -10,9 +10,9 @@ use async_trait::async_trait;
 
 use crate::FtpResult;
 
-#[cfg(feature = "async-native-tls")]
+#[cfg(feature = "async-native-tls-tokio")]
 mod native_tls;
-#[cfg(feature = "async-native-tls")]
+#[cfg(feature = "async-native-tls-tokio")]
 pub use self::native_tls::{AsyncNativeTlsConnector, AsyncNativeTlsStream};
 
 #[cfg(feature = "tokio-rustls")]
